+++ rye/src/bootstrap.rs
@@ -9,16 +9,18 @@ use anyhow::{anyhow, bail, Context, Error};
 use console::style;
 use indicatif::{ProgressBar, ProgressStyle};
 use once_cell::sync::Lazy;
+use tempfile::tempdir_in;
 
 use crate::config::Config;
 use crate::piptools::LATEST_PIP;
 use crate::platform::{
-    get_app_dir, get_canonical_py_path, get_toolchain_python_bin, list_known_toolchains,
+    get_app_dir, get_canonical_py_path, get_python_bin_within, get_toolchain_python_bin,
+    list_known_toolchains,
 };
 use crate::pyproject::latest_available_python_version;
 use crate::sources::py::{get_download_url, PythonVersion, PythonVersionRequest};
 use crate::utils::{check_checksum, symlink_file, unpack_archive, CommandOutput, IoPathContext};
-use crate::uv::Uv;
+use crate::uv::UvBuilder;
 
 /// this is the target version that we want to fetch
 pub const SELF_PYTHON_TARGET_VERSION: PythonVersionRequest = PythonVersionRequest {
@@ -31,11 +33,11 @@ pub const SELF_PYTHON_TARGET_VERSION: PythonVersionRequest = PythonVersionReques
     suffix: None,
 };
 
-const SELF_VERSION: u64 = 14;
+const SELF_VERSION: u64 = 17;
 
 const SELF_REQUIREMENTS: &str = r#"
-build==1.0.3
-certifi==2023.11.17
+build==1.1.1
+certifi==2024.2.2
 charset-normalizer==3.3.2
 click==8.1.7
 distlib==0.3.8
@@ -50,7 +52,7 @@ twine==4.0.2
 unearth==0.14.0
 urllib3==2.0.7
 virtualenv==20.25.0
-ruff==0.2.2
+ruff==0.3.0
 "#;
 
 static FORCED_TO_UPDATE: AtomicBool = AtomicBool::new(false);
@@ -64,6 +66,28 @@ fn is_up_to_date() -> bool {
     *UP_TO_UPDATE || FORCED_TO_UPDATE.load(atomic::Ordering::Relaxed)
 }
 
+#[derive(Debug, Clone)]
+pub(crate) enum SelfVenvStatus {
+    NotUpToDate,
+    DoesNotExist,
+}
+
+/// Get self venv path and check if it exists and is up to date
+pub fn get_self_venv_status() -> Result<PathBuf, (PathBuf, SelfVenvStatus)> {
+    let app_dir = get_app_dir();
+    let venv_dir = app_dir.join("self");
+
+    if venv_dir.is_dir() {
+        if is_up_to_date() {
+            Ok(venv_dir)
+        } else {
+            Err((venv_dir, SelfVenvStatus::NotUpToDate))
+        }
+    } else {
+        Err((venv_dir, SelfVenvStatus::DoesNotExist))
+    }
+}
+
 /// Bootstraps the venv for rye itself
 pub fn ensure_self_venv(output: CommandOutput) -> Result<PathBuf, Error> {
     ensure_self_venv_with_toolchain(output, None)
@@ -75,26 +99,31 @@ pub fn ensure_self_venv_with_toolchain(
     toolchain_version_request: Option<PythonVersionRequest>,
 ) -> Result<PathBuf, Error> {
     let app_dir = get_app_dir();
-    let venv_dir = app_dir.join("self");
 
-    if venv_dir.is_dir() {
-        if is_up_to_date() {
-            return Ok(venv_dir);
-        } else {
-            if output != CommandOutput::Quiet {
-                echo!("Detected outdated rye internals. Refreshing");
-            }
+    let venv_dir = match get_self_venv_status() {
+        Ok(venv_dir) => return Ok(venv_dir),
+        Err((venv_dir, SelfVenvStatus::DoesNotExist)) => venv_dir,
+        Err((venv_dir, SelfVenvStatus::NotUpToDate)) => {
+            echo!(if output, "Detected outdated rye internals. Refreshing");
             fs::remove_dir_all(&venv_dir)
                 .path_context(&venv_dir, "could not remove self-venv for update")?;
-        }
-    }
 
-    if output != CommandOutput::Quiet {
-        echo!("Bootstrapping rye internals");
-    }
+            let pip_tools_dir = app_dir.join("pip-tools");
+            if pip_tools_dir.is_dir() {
+                fs::remove_dir_all(&pip_tools_dir)
+                    .context("could not remove pip-tools for update")?;
+            }
+
+            venv_dir
+        }
+    };
+
+    echo!(if output, "Bootstrapping rye internals");
 
     // Ensure we have uv
-    let uv = Uv::ensure_exists(CommandOutput::Quiet)?;
+    let uv = UvBuilder::new()
+        .with_output(CommandOutput::Quiet)
+        .ensure_exists()?;
 
     let version = match toolchain_version_request {
         Some(ref version_request) => ensure_specific_self_toolchain(output, version_request)
@@ -269,15 +298,17 @@ fn ensure_latest_self_toolchain(output: CommandOutput) -> Result<PythonVersion, 
         .into_iter()
         .max()
     {
-        if output != CommandOutput::Quiet {
-            echo!(
-                "Found a compatible Python version: {}",
-                style(&version).cyan()
-            );
-        }
+        echo!(
+            if output,
+            "Found a compatible Python version: {}",
+            style(&version).cyan()
+        );
         Ok(version)
     } else {
-        fetch(&SELF_PYTHON_TARGET_VERSION, output, false)
+        fetch(
+            &SELF_PYTHON_TARGET_VERSION,
+            FetchOptions::with_output(output),
+        )
     }
 }
 
@@ -295,100 +326,185 @@ fn ensure_specific_self_toolchain(
         );
     }
     if !get_toolchain_python_bin(&toolchain_version)?.is_file() {
-        if output != CommandOutput::Quiet {
-            echo!(
-                "Fetching requested internal toolchain '{}'",
-                toolchain_version
-            );
-        }
-        fetch(&toolchain_version.into(), output, false)
+        echo!(
+            if output,
+            "Fetching requested internal toolchain '{}'",
+            toolchain_version
+        );
+        fetch(&toolchain_version.into(), FetchOptions::with_output(output))
     } else {
-        if output != CommandOutput::Quiet {
-            echo!(
-                "Found a compatible Python version: {}",
-                style(&toolchain_version).cyan()
-            );
-        }
+        echo!(
+            if output,
+            "Found a compatible Python version: {}",
+            style(&toolchain_version).cyan()
+        );
         Ok(toolchain_version)
+    }
+}
+
+/// Fetches a python installer.
+pub struct FetchOptions {
+    /// How verbose should the sync be?
+    pub output: CommandOutput,
+    /// Forces re-downloads even if they are already there.
+    pub force: bool,
+    /// Causes a fetch into a non standard location.
+    pub target_path: Option<PathBuf>,
+    /// Include build info (overrides configured default).
+    pub build_info: Option<bool>,
+}
+
+impl FetchOptions {
+    /// Basic fetch options.
+    pub fn with_output(output: CommandOutput) -> FetchOptions {
+        FetchOptions {
+            output,
+            ..Default::default()
+        }
+    }
+}
+
+impl Default for FetchOptions {
+    fn default() -> Self {
+        Self {
+            output: CommandOutput::Normal,
+            force: false,
+            target_path: None,
+            build_info: None,
+        }
     }
 }
 
 /// Fetches a version if missing.
 pub fn fetch(
     version: &PythonVersionRequest,
-    output: CommandOutput,
-    force: bool,
+    options: FetchOptions,
 ) -> Result<PythonVersion, Error> {
-    if let Ok(version) = PythonVersion::try_from(version.clone()) {
-        let py_bin = get_toolchain_python_bin(&version)?;
-        if !force && py_bin.is_file() {
-            if output == CommandOutput::Verbose {
-                echo!("Python version already downloaded. Skipping.");
+    // Check if there is registered toolchain that matches the request
+    if options.target_path.is_none() {
+        if let Ok(version) = PythonVersion::try_from(version.clone()) {
+            let py_bin = get_toolchain_python_bin(&version)?;
+            if !options.force && py_bin.is_file() {
+                echo!(if verbose options.output, "Python version already downloaded. Skipping.");
+                return Ok(version);
             }
-            return Ok(version);
         }
     }
-
     let (version, url, sha256) = match get_download_url(version) {
         Some(result) => result,
         None => bail!("unknown version {}", version),
     };
 
-    let target_dir = get_canonical_py_path(&version)?;
-    let target_py_bin = get_toolchain_python_bin(&version)?;
-    if output == CommandOutput::Verbose {
-        echo!("target dir: {}", target_dir.display());
-    }
-    if target_dir.is_dir() && target_py_bin.is_file() {
-        if !force {
-            if output == CommandOutput::Verbose {
-                echo!("Python version already downloaded. Skipping.");
+    let target_dir = match options.target_path {
+        Some(ref target_dir) => {
+            if target_dir.is_file() {
+                bail!("target directory '{}' is a file", target_dir.display());
             }
-            return Ok(version);
+            echo!(if options.output, "Downloading to '{}'", target_dir.display());
+            if target_dir.is_dir() {
+                if options.force {
+                    // Refuse to remove the target directory if it's not empty and not a python installation
+                    if target_dir.read_dir()?.next().is_some()
+                        && !get_python_bin_within(target_dir).exists()
+                    {
+                        bail!(
+                            "target directory '{}' exists and is not a Python installation",
+                            target_dir.display()
+                        );
+                    }
+                    fs::remove_dir_all(target_dir)
+                        .path_context(target_dir, "could not remove target directory")?;
+                } else {
+                    bail!("target directory '{}' exists", target_dir.display());
+                }
+            }
+            Cow::Borrowed(target_dir.as_path())
         }
-        if output != CommandOutput::Quiet {
-            echo!("Removing the existing Python version");
+        None => {
+            let target_dir = get_canonical_py_path(&version)?;
+            let target_py_bin = get_toolchain_python_bin(&version)?;
+            if target_py_bin.is_file() {
+                if !options.force {
+                    echo!(if verbose options.output, "Python version already downloaded. Skipping.");
+                    return Ok(version);
+                }
+                echo!(if options.output, "Removing the existing Python version");
+                fs::remove_dir_all(&target_dir).with_context(|| {
+                    format!("failed to remove target folder {}", target_dir.display())
+                })?;
+            }
+            echo!(if verbose options.output, "target dir: {}", target_dir.display());
+            Cow::Owned(target_dir)
         }
-        fs::remove_dir_all(&target_dir)
-            .with_context(|| format!("failed to remove target folder {}", target_dir.display()))?;
-    }
+    };
 
-    fs::create_dir_all(&target_dir).path_context(&target_dir, "failed to create target folder")?;
-
-    if output == CommandOutput::Verbose {
-        echo!("download url: {}", url);
-    }
-    if output != CommandOutput::Quiet {
-        echo!("{} {}", style("Downloading").cyan(), version);
-    }
-    let archive_buffer = download_url(url, output)?;
+    echo!(if verbose options.output, "download url: {}", url);
+    echo!(if options.output, "{} {}", style("Downloading").cyan(), version);
+    let archive_buffer = download_url(url, options.output)?;
 
     if let Some(sha256) = sha256 {
-        if output != CommandOutput::Quiet {
-            echo!("{} {}", style("Checking").cyan(), "checksum");
-        }
+        echo!(if options.output, "{} {}", style("Checking").cyan(), "checksum");
         check_checksum(&archive_buffer, sha256)
             .with_context(|| format!("Checksum check of {} failed", &url))?;
-    } else if output != CommandOutput::Quiet {
-        echo!("Checksum check skipped (no hash available)");
+    } else {
+        echo!(if options.output, "Checksum check skipped (no hash available)");
     }
 
-    if output != CommandOutput::Quiet {
-        echo!("{}", style("Unpacking").cyan());
+    echo!(if options.output, "{}", style("Unpacking").cyan());
+
+    let parent = target_dir
+        .parent()
+        .ok_or_else(|| anyhow!("cannot unpack to root"))?;
+    if !parent.exists() {
+        fs::create_dir_all(parent).path_context(&target_dir, "failed to create target folder")?;
     }
-    unpack_archive(&archive_buffer, &target_dir, 1).with_context(|| {
+
+    let with_build_info = options
+        .build_info
+        .unwrap_or_else(|| Config::current().fetch_with_build_info());
+    let temp_dir = tempdir_in(parent).context("temporary unpack location")?;
+
+    unpack_archive(&archive_buffer, temp_dir.path(), 1).with_context(|| {
         format!(
             "unpacking of downloaded tarball {} to '{}' failed",
             &url,
-            target_dir.display()
+            temp_dir.path().display(),
         )
     })?;
 
-    if output != CommandOutput::Quiet {
-        echo!("{} {}", style("Downloaded").green(), version);
+    // if we want to retain build infos or the installation has no build infos, then move
+    // the folder into the permanent location
+    if with_build_info || !installation_has_build_info(temp_dir.path()) {
+        let temp_dir = temp_dir.into_path();
+        fs::rename(&temp_dir, &target_dir).map_err(|err| {
+            fs::remove_dir_all(&temp_dir).ok();
+            err
+        })
+
+    // otherwise move the contents of the `install` folder over.
+    } else {
+        fs::rename(temp_dir.path().join("install"), &target_dir)
     }
+    .path_context(&target_dir, "unable to persist download")?;
+
+    echo!(if options.output, "{} {}", style("Downloaded").green(), version);
 
     Ok(version)
+}
+
+fn installation_has_build_info(p: &Path) -> bool {
+    let mut has_install = false;
+    let mut has_build = false;
+    if let Ok(dir) = p.read_dir() {
+        for entry in dir.flatten() {
+            match entry.file_name().to_str() {
+                Some("install") => has_install = true,
+                Some("build") => has_build = true,
+                _ => {}
+            }
+        }
+    }
+    has_install && has_build
 }
 
 pub fn download_url(url: &str, output: CommandOutput) -> Result<Vec<u8>, Error> {