+++ CHANGELOG.md
@@ -3,16 +3,102 @@
 This file contains tracks the changes landing in Rye.  It includes changes
 that were not yet released.
 
-## 0.28.0
+## 0.33.0
 
 _Unreleased_
+
+<!-- released start -->
+
+## 0.32.0
+
+Released on 2024-03-29
+
+- Update uv to 0.1.26.  #924
+
+- Always create `.gitignore` file in `rye init`.  #919
+
+- Prevent `rye fetch --force` from removing a target directory that is not a Python installation.  #921
+
+- `rye list` always prints the currently installed packages even this project is not managed by Rye.  #940
+
+- Fix error on using -v or -q with `rye fmt` or `rye lint`. #959
+
+- Fix rye fetch detection of registered toolchain.  #931
+
+- Ignore build-system configuration for virtual projects.  #929
+
+## 0.31.0
+
+Released on 2024-03-22
+
+- Update uv to 0.1.23.  #916
+
+- Allow `rye publish` working outside of project.  #910
+
+- `rye test --quiet` no longer implies `--no-capture`. #915
+
+- Rye now can be used to fetch Python installations even when not using Rye
+  and build infos are no longer included by default.  This means that rather
+  than having interpreters at `~/.rye/py/cpython@3.11.1/install/bin/python3`
+  it will now reside at `~/.rye/py/cpython@3.11.1/bin/python3`.  #917
+
+- Installer now recommends `uv` over `pip-tools`.  #918
+
+## 0.30.0
+
+Released on 2024-03-19
+
+- Update uv to 0.1.21.  #884, #890, #904
+
+- Fix incorrect flag passing of `rye test` `-q` and `-v`.  #880
+
+- Rye now loads `.env` files.  This applies both for Rye's own
+  use of environment variables but also to scripts launched via
+  `run`.  #894
+
+- Fix `rye add m --path ./m` causing a panic on windows.  #897
+
+## 0.29.0
+
+Released on 2024-03-11
+
+- Updated to `uv` 0.1.17.  #850, #867
+
+- Trap panics and silence bad pipe errors.  #862
+
+- Updating `rye` will now also ensure that the self-venv is updated.  Previously
+  this was deferred until the next `sync`.  #863
+
+- The `self update` command now accepts `--branch`.  #864
+
+- Fixed an issue that caused pip-tools to not update.  #865
+
+- Updates `build` and `certifi`.  #866
+
+## 0.28.0
+
+Released on 2024-03-07
+
+- `--skip-existing` is now available with Rye's `publish` command. #831
+
+- Bumped `uv` to 0.1.15.  #760, #820, #837
+
+- Bumped `ruff` to 0.3.0.  #821
 
 - The `init` command now generates a script with the name of the
   project rather than `hello`.  #801
 
 - Retain markers when adding dependencies with features when uv is used.  #807
 
-<!-- released start -->
+- Fixed a bug that caused repeated syncs not to recall all previous options.  #830
+
+- Report `self-python` version in `--version`.  #843
+
+- Fixes a bug where `rye config` would not create the `RYE_HOME` folder if needed.  #844
+
+- `rye add` now retains version and URL for the requirements when `uv` is used.  #846
+
+- Added a `rye test` command which invokes `pytest`.  #847
 
 ## 0.27.0